+++ packages/wallet-sdk/src/sign/SignerConfigurator.ts
@@ -1,8 +1,8 @@
 import { SCWSigner } from './scw/SCWSigner';
-import { PopUpCommunicator } from './scw/transport/PopUpCommunicator';
 import { Signer } from './SignerInterface';
 import { SignRequestHandlerListener } from './UpdateListenerInterface';
 import { WLSigner } from './walletlink/WLSigner';
+import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
 import { standardErrors } from ':core/error';
 import { SignerType, WalletLinkConfigEventType } from ':core/message/ConfigMessage';
 import { ScopedLocalStorage } from ':core/storage/ScopedLocalStorage';
